pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goal-drift-gym"
version = "0.0.1"
description = "Gym-style harness for measuring goal drift in decision-making agents"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Tyler Crosse"}]
license = {text = "MIT"}
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "matplotlib>=3.7",
    "tqdm>=4.66",
]

[tool.setuptools]
package-dir = {"" = "goal_drift_gym"}

[tool.setuptools.packages.find]
where = ["goal_drift_gym"]
